refactor: move airdrop FE related code to own package
