+++ src/frontend/src/airdrop/components/airdrop/AirdropButton.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
-	import IconCelebrate from '$lib/components/icons/IconCelebrate.svelte';
+	import IconCelebrate from '$airdrop/components/icons/IconCelebrate.svelte';
 	import { slide } from 'svelte/transition';
 	import { modalStore } from '$lib/stores/modal.store';
-	import AirdropModal from '$lib/components/airdrop/AirdropModal.svelte';
+	import AirdropModal from '$airdrop/components/airdrop/AirdropModal.svelte';
 	import { nonNullish } from '@dfinity/utils';
-	import { airdropStore } from '$lib/stores/airdrop.store';
+	import { airdropStore } from '$airdrop/stores/airdrop.store';
 	import { quintOut } from 'svelte/easing';
 </script>
