+++ src/frontend/src/airdrop/components/airdrop/AirdropCompleted.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import type { Info } from '$declarations/airdrop/airdrop.did';
 	import type { StaticStep } from '$lib/types/steps';
-	import { countAirdropInvitesRedeemed } from '$lib/utils/airdrop.utils';
+	import { countAirdropInvitesRedeemed } from '$airdrop/utils/airdrop.utils';
 	import StaticSteps from '$lib/components/ui/StaticSteps.svelte';
 	import { AirdropStep } from '$lib/enums/steps';
