+++ src/frontend/src/airdrop/components/icons/IconCelebrate.svelte
