+++ src/frontend/src/airdrop/components/icons/IconShare.svelte
