+++ src/frontend/src/airdrop/types/airdrop.ts
