+++ src/frontend/src/airdrop/utils/airdrop.utils.ts
