+++ src/frontend/src/airdrop/utils/canvas.utils.ts
