+++ src/frontend/src/lib/components/core/Workers.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import AirdropWorker from '$lib/components/airdrop/AirdropWorker.svelte';
+	import AirdropWorker from '$airdrop/components/airdrop/AirdropWorker.svelte';
 	import ExchangeWorker from '$lib/components/exchange/ExchangeWorker.svelte';
 </script>
